Extract DrmFakeDirManager to a separate helper
